requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
python-dotenv==1.0.0
python-slugify==8.0.1
requests==2.28.2
text-unidecode==1.3
urllib3==1.26.15
